Some cleanup & improve images
