+++ src/app/components/FrostedGlassPromo/lazy.jsx
@@ -1,0 +1,3 @@
+import loadable from '@loadable/component';
+
+export default loadable(() => import('.'));
